CSV Forcings: Move implementation to a .cpp file
