+++ src/forcing/CMakeLists.txt
@@ -21,6 +21,7 @@ target_link_libraries(forcing PUBLIC
 )
 
 target_sources(forcing PRIVATE "${CMAKE_CURRENT_LIST_DIR}/NullForcingProvider.cpp")
+target_sources(forcing PRIVATE "${CMAKE_CURRENT_LIST_DIR}/CsvPerFeatureForcingProvider.cpp")
 
 if(NGEN_WITH_NETCDF)
     target_sources(forcing PRIVATE "${CMAKE_CURRENT_LIST_DIR}/NetCDFPerFeatureDataProvider.cpp")
